requirement.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
translate==3.6.1
urllib3==2.0.3
yarl==1.9.2
